+++ engine/apps/public_api/views/users.py
@@ -1,3 +1,4 @@
+from django_filters import rest_framework as filters
 from rest_framework.decorators import action
 from rest_framework.exceptions import NotFound
 from rest_framework.permissions import IsAuthenticated
@@ -16,6 +17,20 @@ from common.api_helpers.paginators import HundredPageSizePaginator
 from common.constants.role import Role
 
 
+class UserFilter(filters.FilterSet):
+    """
+    https://django-filter.readthedocs.io/en/master/guide/rest_framework.html
+    """
+
+    email = filters.CharFilter(field_name="email", lookup_expr="iexact")
+    roles = filters.MultipleChoiceFilter(field_name="role", choices=Role.choices())
+    username = filters.CharFilter(field_name="username", lookup_expr="iexact")
+
+    class Meta:
+        model = User
+        fields = ["email", "roles", "username"]
+
+
 class UserView(RateLimitHeadersMixin, ShortSerializerMixin, ReadOnlyModelViewSet):
     authentication_classes = (ApiTokenAuthentication,)
     permission_classes = (IsAuthenticated,)
@@ -25,23 +40,17 @@ class UserView(RateLimitHeadersMixin, ShortSerializerMixin, ReadOnlyModelViewSet
 
     serializer_class = UserSerializer
     short_serializer_class = FastUserSerializer
+    filterset_class = UserFilter
+    filter_backends = (filters.DjangoFilterBackend,)
 
     throttle_classes = [UserThrottle]
 
     def get_queryset(self):
-        username = self.request.query_params.get("username")
-        email = self.request.query_params.get("email")
         is_short_request = self.request.query_params.get("short", "false") == "true"
-        queryset = self.request.auth.organization.users.filter(role__in=[Role.ADMIN, Role.EDITOR]).distinct()
-
-        if username is not None:
-            queryset = queryset.filter(username=username)
-
-        if email is not None:
-            queryset = queryset.filter(email=email)
-
+        queryset = self.request.auth.organization.users.all()
         if not is_short_request:
             queryset = self.serializer_class.setup_eager_loading(queryset)
+        queryset = self.filter_queryset(queryset)
         return queryset.order_by("id")
 
     def get_object(self):