+++ grafana-plugin/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Change Log
 
+## 1.0.2 (2022-06-17)
+
+- Fix Grafana Alerting integration to handle API changes in Grafana 9
+- Improve public API endpoint for outgoing webhooks (/actions) by adding ability to create, update and delete
+
+## 1.0.0 (2022-06-14)
+
+- First Public Release
+
 ## 0.0.71 (2022-06-06)
 
-- Initial Release
+- Initial Commit Release